% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta.R
\name{beta}
\alias{beta}
\title{Beta growth model}
\usage{
beta(data, Strain)
}
\arguments{
\item{data}{data frame containing equally/unequally spaced time 
points}

\item{Strain}{strain}

\item{MouseID}{mouseID}
}
\value{
fitted model?
}
\description{
Fit a beta growth model (with four parameters) to the data
}
\examples{
beta(data = df_equal, MouseID = "LL1-1_BalbcJ", Strain = "BALB/cJ")
}
