% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_data.R
\name{filter_data}
\alias{filter_data}
\title{Filter data}
\usage{
filter_data(data, threshold)
}
\arguments{
\item{data}{data frame containing unequally spaced time 
points}

\item{threshold}{minimum cutoff for unequally spaced time points}
}
\value{
filtered data set
}
\description{
Remove mice with too few data points
}
\examples{
filter_data(data = df_unequal, threshold = 5)
}
