% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_bayes_growth_model.R
\name{fit_bayes_growth_model}
\alias{fit_bayes_growth_model}
\title{Fit growth model in a Bayesian framework}
\usage{
fit_bayes_growth_model(data, Strain, type)
}
\arguments{
\item{data}{data frame containing equally/unequally spaced time 
points}

\item{Strain}{Strain of interest}

\item{type}{type of growth curve}
}
\value{
Estimated growth curve parameters
}
\description{
Fit a growth model to the data
}
\examples{
fit_growth_model(data = df_equal, Strain = "C57BL/6NJ", type = "Gompertz")
}
