% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth_model.R
\name{growth_model}
\alias{growth_model}
\title{growth model}
\usage{
growth_model(data, type)
}
\arguments{
\item{data}{data frame containing equally/unequally spaced time 
points}

\item{type}{type of growth curve}
}
\value{
Estimated growth curve parameters
}
\description{
Fit a growth model to the data
}
\examples{
growth_model(data = df_equal, type = "Gompertz")
}
