% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_data.R
\name{merge_data}
\alias{merge_data}
\title{Load and preprocess data}
\usage{
merge_data(df1, df2)
}
\arguments{
\item{df1}{data frame containing growth parameters}

\item{df2}{data frame containing grooming and open field data}
}
\value{
a data frame containing growth parameters and other traits #' merged on common MouseIDs
}
\description{
Loads the data and preprocess the features.
}
\examples{
merge_data(df1, df2)
}
