% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fboli_plots.R
\name{plot_data}
\alias{plot_data}
\title{Exploratory plot}
\usage{
plot_data(data)
}
\arguments{
\item{data}{data frame containing equally/unequally spaced time 
points}
}
\value{
plot
}
\description{
Plots the raw data (both equally and unequally spaced time points)
}
\examples{
plot_data(data = df_equal)
}
