% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fboli_plots.R
\name{plot_growth_comparison3}
\alias{plot_growth_comparison3}
\title{Plot simulated growth curves}
\usage{
plot_growth_comparison3(data, type)
}
\arguments{
\item{data}{an aggregated data frame containing estimated growth #' parameters for each strain}

\item{type}{type of plot}
}
\value{
a plot containing simulated growth curves for all strains
}
\description{
Plots the simulated growth curves using estimated growth 
parameters for comparison 'faceted' across all strains
}
\examples{
plot_growth_comparison(data = df_growth, type = "Strain")
}
