% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fboli_plots.R
\name{plot_growth_model}
\alias{plot_growth_model}
\title{Plot fitted growth curves on raw data}
\usage{
plot_growth_model(data, Strain, shrinkage)
}
\arguments{
\item{data}{data frame containing equally/unequally spaced time 
points}

\item{Strain}{strain}
}
\value{
a list containing the plot and a data frame with estimated #' growth parameters
}
\description{
Plots the fitted growth curve for all animals belonging to a strain
}
\examples{
plot_growth_model(data = df_unequal, Strain = "C57BL/6NJ")
}
