% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fboli_plots.R
\name{plot_scatter_strain}
\alias{plot_scatter_strain}
\title{Plot open field trait vs estimated growth parameter}
\usage{
plot_scatter_strain(data, of_trait, growth_param)
}
\arguments{
\item{data}{data frame containing estimated growth parameters 
and other open field and grooming traits}

\item{of_trait}{open field trait}

\item{growth_param}{estimated growth curve parameter}
}
\value{
a scatter plot at the strain level
}
\description{
A scatter plot of an open field trait vs estimated growth 
parameter at the strain level
}
\examples{
plot_scatter_strain(data = df_merge, of_trait = 
"bin_avg_55.periphery_time_secs", growth_param = "Asym")
}
