% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fboli_plots.R
\name{plot_scatter_strain_all}
\alias{plot_scatter_strain_all}
\title{Plot all open field traits vs estimated growth parameter}
\usage{
plot_scatter_strain_all(data, growth_param)
}
\arguments{
\item{data}{data frame containing estimated growth parameters 
and other open field and grooming traits}

\item{growth_param}{estimated growth curve parameter}

\item{of_trait}{open field trait}
}
\value{
a scatter plot at the strain level
}
\description{
A scatter plot of all open field traits vs estimated growth 
parameter at the strain level
}
\examples{
plot_scatter_strain_all(data = df_merge, growth_param = "k")
}
