% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fboli_plots.R
\name{plot_signif_corr}
\alias{plot_signif_corr}
\title{Plot grooming paper like plots}
\usage{
plot_signif_corr(data, of_trait, growth_param)
}
\arguments{
\item{data}{data frame containing estimated growth parameters 
and other open field and grooming traits}

\item{of_trait}{open field trait}

\item{growth_param}{estimated growth curve parameter}
}
\value{
a plot highlighting Strains with significant correlations
}
\description{
Compare growth parameters with open field traits across strains
and highlight strains for which correlations are signficant
}
\examples{
plot_signif_corr(data, of_trait = "bin_avg_55.periphery_time_secs"
,growth_param = "Asym")
}
