% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fboli_plots.R
\name{plot_strain_animal}
\alias{plot_strain_animal}
\title{Exploratory plot}
\usage{
plot_strain_animal(data, MouseID, Strain)
}
\arguments{
\item{data}{data frame containing equally/unequally spaced time 
points}

\item{MouseID}{mouseID}

\item{Strain}{strain}
}
\value{
plot
}
\description{
Plots the raw data (both equally and unequally spaced time points)
for a single animal/strain
}
\examples{
plot_strain_animal(data = df_equal, MouseID = "LL1-1_BalbcJ",
Strain = NULL)
plot_strain_animal(data = df_equal, MouseID = NULL, Strain = 
Strain = "BALB/cJ")
}
