% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_data}
\alias{load_data}
\title{Load and preprocess data}
\usage{
load_data(csv)
}
\arguments{
\item{csv}{a csv file containing data per stride}
}
\value{
a list containing two data frames
1) df_equal: equally spaced time points
2) df_unequal: un-equally spaced time points
}
\description{
Loads the data and preprocess the features.
}
\examples{
load_data(data = "fecalboli.csv")
}
